remove /api/reports
