+++ reports/src/main/resources/templates/create_report.html
@@ -43,7 +43,7 @@
                                 Filter Example
                             </div>
                             <p>
-                                tagKey == 'CUSTOMER' && (tagValue == 'Demo' || tagValue == 'Test')
+                                tagKey == 'CUSTOMER' && (tagValue == 'Demo' || tagValue == 'Test')<br>
                                 tagKey == 'DOMAIN' && !tagValue.isEmpty()
                             </p>
                         </div>
